requirements.txt
# Minimal runtime packages (install with: pip install -r requirements.txt)
numpy
opencv-python
Pillow
ultralytics
streamlit
